--- app/worker/items.js
@@ -1,380 +1,0 @@
-const fs = require('fs');
-const DatabaseManager = require('../lib/DatabaseManager');
-
-const dbManager = new DatabaseManager({
-  mongoURL: process.env.MONGO_URL,
-});
-
-const CACHE_DIR = 'app/worker/cache';
-const ARMOR_DIR = `${CACHE_DIR}/armor`;
-// const WEAPON_DIR = `${CACHE_DIR}/weapon`;
-const BUCKET_DIR = `${CACHE_DIR}/bucket`;
-const MOD_DIR = `${CACHE_DIR}/mod`;
-const STAT_DIR = `${CACHE_DIR}/stat`;
-
-const HASH_SOCKET_STAT = 4076485920;
-const HASH_FAT_SOCKET_STAT = 3497077129;
-const HASH_STAT_DEFENSE = 3897883278;
-const HASH_STAT_POWER = 1935470627;
-const HASH_STAT_MOBILITY = 2996146975;
-const HASH_STAT_RESILIENCE = 392767087;
-const HASH_STAT_RECOVERY = 1943323491;
-
-const BUNGIE_ROOT = 'https://www.bungie.net';
-
-const stats = [];
-const mods = [];
-const buckets = [];
-const armors = [];
-
-/**
- * Read the stats files and build Stat objects
- * @param {String} file
- * @param {String} lang
- */
-function readStatFile(file, lang) {
-  return new Promise((resolve, reject) => {
-    fs.readFile(`${STAT_DIR}/${file}`, (err, data) => {
-      if (err) reject(err);
-      const json = JSON.parse(data);
-      Object.keys(json).forEach((id) => {
-        const stat = json[id];
-        if (stat.displayProperties.name !== '') {
-          const existingStat = stats.find(element => element.hash === id);
-          if (existingStat) {
-            existingStat.names[lang] = stat.displayProperties.name;
-            existingStat.descriptions[lang] = stat.displayProperties.description;
-          } else {
-            const obj = {
-              hash: id,
-              names: {
-                [lang]: stat.displayProperties.name,
-              },
-              descriptions: {
-                [lang]: stat.displayProperties.description,
-              },
-            };
-            if (stat.displayProperties.hasIcon) {
-              obj.img = `${BUNGIE_ROOT}${stat.displayProperties.icon}`;
-            }
-            stats.push(obj);
-          }
-        }
-      });
-      resolve();
-    });
-  });
-}
-
-/**
- * Read the mods files and build Mod objects
- * @param {String} file
- * @param {String} lang
- */
-function readModFile(file, lang) {
-  return new Promise((resolve, reject) => {
-    fs.readFile(`${MOD_DIR}/${file}`, (err, data) => {
-      if (err) reject(err);
-      const json = JSON.parse(data);
-      Object.keys(json).forEach(async (id) => {
-        const mod = json[id];
-        if (mod.displayProperties.name !== '') {
-          const existingMod = mods.find(element => element.hash === id);
-          if (existingMod) {
-            existingMod.names[lang] = mod.displayProperties.name;
-            existingMod.descriptions[lang] = mod.displayProperties.description;
-          } else {
-            const obj = {
-              hash: id,
-              names: {
-                [lang]: mod.displayProperties.name,
-              },
-              descriptions: {
-                [lang]: mod.displayProperties.description,
-              },
-              mobility: 0,
-              resilience: 0,
-              recovery: 0,
-            };
-            if (mod.displayProperties.hasIcon) {
-              obj.img = `${BUNGIE_ROOT}${mod.displayProperties.icon}`;
-            }
-            if (mod.plug.plugCategoryIdentifier !== undefined) {
-              obj.type = mod.plug.plugCategoryIdentifier;
-            }
-            const promises = [];
-            if (mod.investmentStats !== undefined) {
-              mod.investmentStats.forEach((stat) => {
-                promises.push(new Promise((elResolve) => {
-                  switch (stat.statTypeHash) {
-                    case HASH_STAT_MOBILITY: {
-                      obj.mobility += stat.value;
-                      break;
-                    }
-                    case HASH_STAT_RESILIENCE: {
-                      obj.resilience += stat.value;
-                      break;
-                    }
-                    case HASH_STAT_RECOVERY: {
-                      obj.recovery += stat.value;
-                      break;
-                    }
-                    default:
-                      break;
-                  }
-                  elResolve();
-                }));
-              });
-            }
-            await Promise.all(promises);
-            mods.push(obj);
-          }
-        }
-      });
-      resolve();
-    });
-  });
-}
-
-/**
- * Read the buckets files and build Bucket objects
- * @param {String} file
- * @param {String} lang
- */
-function readBucketFile(file, lang) {
-  return new Promise((resolve, reject) => {
-    fs.readFile(`${BUCKET_DIR}/${file}`, (err, data) => {
-      if (err) reject(err);
-      const json = JSON.parse(data);
-      Object.keys(json).forEach((id) => {
-        const bucket = json[id];
-        if (bucket.displayProperties.name !== '') {
-          const existingBucket = buckets.find(element => element.hash === id);
-          if (existingBucket) {
-            existingBucket.names[lang] = bucket.displayProperties.name;
-            existingBucket.descriptions[lang] = bucket.displayProperties.description;
-          } else {
-            const obj = {
-              hash: id,
-              names: {
-                [lang]: bucket.displayProperties.name,
-              },
-              descriptions: {
-                [lang]: bucket.displayProperties.description,
-              },
-            };
-            buckets.push(obj);
-          }
-        }
-      });
-      resolve();
-    });
-  });
-}
-
-/**
- * Read the armors files and build Armor objects
- * @param {String} file
- * @param {String} lang
- */
-function readArmorFile(file, lang) {
-  return new Promise((resolve, reject) => {
-    fs.readFile(`${ARMOR_DIR}/${file}`, (err, data) => {
-      if (err) reject(err);
-      const json = JSON.parse(data);
-      Object.keys(json).forEach((id) => {
-        const armor = json[id];
-        if (armor.displayProperties.name !== '' && armor.inventory.tierType > 4) {
-          const existingArmor = armors.find(element => element.hash === id);
-          if (existingArmor) {
-            existingArmor.names[lang] = armor.displayProperties.name;
-            existingArmor.descriptions[lang] = armor.displayProperties.description;
-          } else {
-            const obj = {
-              hash: id,
-              names: {
-                [lang]: armor.displayProperties.name,
-              },
-              descriptions: {
-                [lang]: armor.displayProperties.description,
-              },
-              img: armor.displayProperties.hasIcon ? `${BUNGIE_ROOT}${armor.displayProperties.icon}` : '',
-              screenshot: armor.screenshot !== undefined ? `${BUNGIE_ROOT}${armor.screenshot}` : '',
-              tier: armor.inventory.tierType,
-              defense: 0,
-              power: 0,
-              mobility: 0,
-              resilience: 0,
-              recovery: 0,
-              mods: [],
-            };
-            if (armor.sourceData !== undefined && armor.sourceData.sources !== undefined) {
-              armor.sourceData.sources.forEach((level) => {
-                if (level.computedStats[HASH_STAT_DEFENSE] !== undefined) {
-                  if (level.computedStats[HASH_STAT_DEFENSE].value > obj.defense) {
-                    obj.defense = level.computedStats[HASH_STAT_DEFENSE].value;
-                  }
-                }
-                if (level.computedStats[HASH_STAT_POWER] !== undefined) {
-                  if (level.computedStats[HASH_STAT_POWER].value > obj.power) {
-                    obj.power = level.computedStats[HASH_STAT_POWER].value;
-                  }
-                }
-                if (level.computedStats[HASH_STAT_MOBILITY] !== undefined) {
-                  if (level.computedStats[HASH_STAT_MOBILITY].value > obj.mobility) {
-                    obj.mobility = level.computedStats[HASH_STAT_MOBILITY].value;
-                  }
-                }
-                if (level.computedStats[HASH_STAT_RESILIENCE] !== undefined) {
-                  if (level.computedStats[HASH_STAT_RESILIENCE].value > obj.resilience) {
-                    obj.resilience = level.computedStats[HASH_STAT_RESILIENCE].value;
-                  }
-                }
-                if (level.computedStats[HASH_STAT_RECOVERY] !== undefined) {
-                  if (level.computedStats[HASH_STAT_RECOVERY].value > obj.recovery) {
-                    obj.recovery = level.computedStats[HASH_STAT_RECOVERY].value;
-                  }
-                }
-              });
-            }
-            if (armor.sockets !== undefined && armor.sockets.socketEntries) {
-              armor.sockets.socketEntries.forEach((socket) => {
-                if (socket.socketTypeHash === HASH_FAT_SOCKET_STAT && socket.reusablePlugItems !== undefined) {
-                  socket.reusablePlugItems.forEach(async (modHash) => {
-                    try {
-                      const fatMod = await dbManager.findModByHash(modHash.plugItemHash);
-                      obj.mobility += fatMod.mobility;
-                      obj.resilience += fatMod.resilience;
-                      obj.recovery += fatMod.recovery;
-                    } catch (modErr) {
-                      console.log(modErr);
-                    }
-                  });
-                }
-                if (socket.socketTypeHash === HASH_SOCKET_STAT && socket.reusablePlugItems !== undefined) {
-                  socket.reusablePlugItems.forEach(async (modHash) => {
-                    const mod = await dbManager.findModByHash(modHash.plugItemHash);
-                    obj.mods.push(mod._id);
-                  });
-                }
-              });
-            }
-            armors.push(obj);
-          }
-        }
-      });
-      resolve();
-    });
-  });
-}
-
-module.exports = {
-  saveStats: () => new Promise(async (resolve, reject) => {
-    try {
-      await dbManager.connect();
-    } catch (err) {
-      reject(err);
-    }
-    fs.readdir(STAT_DIR, async (err, files) => {
-      if (err) reject(err);
-      else {
-        await dbManager.removeAllStats();
-        const promises = [];
-        files.forEach((file) => {
-          const lang = file.split('.')[0];
-          promises.push(readStatFile(file, lang));
-        });
-        Promise.all(promises).then(async () => {
-          try {
-            await dbManager.saveStats(stats);
-            await dbManager.disconnect();
-            resolve();
-          } catch (pErr) {
-            console.log(`An error occured while saving the stats ${pErr}`);
-          }
-        });
-      }
-    });
-  }),
-  saveMods: () => new Promise(async (resolve, reject) => {
-    try {
-      await dbManager.connect();
-    } catch (err) {
-      reject(err);
-    }
-    fs.readdir(MOD_DIR, async (err, files) => {
-      if (err) reject(err);
-      else {
-        await dbManager.removeAllMods();
-        const promises = [];
-        files.forEach((file) => {
-          const lang = file.split('.')[0];
-          promises.push(readModFile(file, lang));
-        });
-        Promise.all(promises).then(async () => {
-          try {
-            await dbManager.saveMods(mods);
-            await dbManager.disconnect();
-            resolve();
-          } catch (pErr) {
-            console.log(`An error occured while saving the mods ${pErr}`);
-          }
-        });
-      }
-    });
-  }),
-  saveBuckets: () => new Promise(async (resolve, reject) => {
-    try {
-      await dbManager.connect();
-    } catch (err) {
-      reject(err);
-    }
-    fs.readdir(BUCKET_DIR, async (err, files) => {
-      if (err) reject(err);
-      else {
-        await dbManager.removeAllBuckets();
-        const promises = [];
-        files.forEach((file) => {
-          const lang = file.split('.')[0];
-          promises.push(readBucketFile(file, lang));
-        });
-        Promise.all(promises).then(async () => {
-          try {
-            await dbManager.saveBuckets(buckets);
-            await dbManager.disconnect();
-            resolve();
-          } catch (pErr) {
-            console.log(`An error occured while saving the buckets ${pErr}`);
-          }
-        });
-      }
-    });
-  }),
-  saveArmors: () => new Promise(async (resolve, reject) => {
-    try {
-      await dbManager.connect();
-    } catch (err) {
-      reject(err);
-    }
-    fs.readdir(ARMOR_DIR, async (err, files) => {
-      if (err) reject(err);
-      else {
-        await dbManager.removeAllArmors();
-        const promises = [];
-        files.forEach((file) => {
-          const lang = file.split('.')[0];
-          promises.push(readArmorFile(file, lang));
-        });
-        Promise.all(promises).then(async () => {
-          try {
-            await dbManager.saveArmors(armors);
-            await dbManager.disconnect();
-            resolve();
-          } catch (pErr) {
-            console.log(`An error occured while saving the armors ${pErr}`);
-          }
-        });
-      }
-    });
-  }),
-};
